+++ packages/dash-table/src/dash-table/components/FilterFactory.tsx
@@ -57,13 +57,27 @@ export default class FilterFactory {
         updateColumnFilter(map, column, operator, value, setFilter);
     };
 
+    private onToggleChange = (
+        column: IColumn,
+        map: Map<string, SingleColumnSyntaxTree>,
+        operator: FilterLogicalOperator,
+        setFilter: SetFilter,
+        toggleFilterOptions: (column: IColumn) => IColumn,
+        value: any
+    ) => {
+        const newColumn = toggleFilterOptions(column);
+
+        updateColumnFilter(map, newColumn, operator, value, setFilter);
+    };
+
     private filter = memoizerCache<[ColumnId, number]>()(
         (
             column: IColumn,
             index: number,
             map: Map<string, SingleColumnSyntaxTree>,
             operator: FilterLogicalOperator,
-            setFilter: SetFilter
+            setFilter: SetFilter,
+            toggleFilterOptions: (column: IColumn) => IColumn
         ) => {
             const ast = map.get(column.id.toString());
 
@@ -72,6 +86,7 @@ export default class FilterFactory {
                     key={`column-${index}`}
                     className={`dash-filter column-${index}`}
                     columnId={column.id}
+                    filterOptions={column.filter_options}
                     isValid={!ast || ast.isValid}
                     setFilter={this.onChange.bind(
                         this,
@@ -80,6 +95,11 @@ export default class FilterFactory {
                         operator,
                         setFilter
                     )}
+                    // Running into TypeScript binding issues with many parameters..
+                    // bind with no more than 4 params each time.. sigh..
+                    toggleFilterOptions={this.onToggleChange
+                        .bind(this, column, map, operator, setFilter)
+                        .bind(this, toggleFilterOptions, ast && ast.query)}
                     value={ast && ast.query}
                 />
             );
@@ -107,6 +127,7 @@ export default class FilterFactory {
             style_cell_conditional,
             style_filter,
             style_filter_conditional,
+            toggleFilterOptions,
             visibleColumns
         } = this.props;
 
@@ -139,7 +160,8 @@ export default class FilterFactory {
                     index,
                     map,
                     filter_action.operator,
-                    setFilter
+                    setFilter,
+                    toggleFilterOptions
                 );
             },
             visibleColumns