+++ packages/dash-table/src/dash-table/components/Table/props.ts
@@ -38,6 +38,11 @@ export enum ExportHeaders {
     Display = 'display'
 }
 
+export enum FilterCase {
+    Insensitive = 'insensitive',
+    Sensitive = 'sensitive'
+}
+
 export enum SortMode {
     Single = 'single',
     Multi = 'multi'
@@ -47,6 +52,10 @@ export enum TableAction {
     Custom = 'custom',
     Native = 'native',
     None = 'none'
+}
+
+export interface IFilterOptions {
+    case?: FilterCase;
 }
 
 export interface IDerivedData {
@@ -193,6 +202,7 @@ export interface IBaseColumn {
     clearable?: boolean | boolean[] | 'first' | 'last';
     deletable?: boolean | boolean[] | 'first' | 'last';
     editable: boolean;
+    filter_options: IFilterOptions;
     hideable?: boolean | boolean[] | 'first' | 'last';
     renamable?: boolean | boolean[] | 'first' | 'last';
     selectable?: boolean | boolean[] | 'first' | 'last';
@@ -327,6 +337,7 @@ export interface IProps {
     data?: Data;
     editable?: boolean;
     fill_width?: boolean;
+    filter_options?: IFilterOptions;
     filter_query?: string;
     filter_action?: TableAction | IFilterAction;
     hidden_columns?: string[];
@@ -381,6 +392,7 @@ interface IDefaultProps {
     export_format: ExportFormat;
     export_headers: ExportHeaders;
     fill_width: boolean;
+    filter_options?: IFilterOptions;
     filter_query: string;
     filter_action: TableAction;
     fixed_columns: Fixed;
@@ -492,6 +504,7 @@ export interface IFilterFactoryProps {
     style_cell_conditional: Cells;
     style_filter: Style;
     style_filter_conditional: BasicFilters;
+    toggleFilterOptions: (column: IColumn) => IColumn;
     visibleColumns: Columns;
 }