+++ packages/dash-table/src/dash-table/dash/Sanitizer.ts
@@ -16,7 +16,9 @@ import {
     ExportHeaders,
     IFilterAction,
     FilterLogicalOperator,
-    SelectedCells
+    SelectedCells,
+    FilterCase,
+    IFilterOptions
 } from 'dash-table/components/Table/props';
 import headerRows from 'dash-table/derived/header/headerRows';
 import resolveFlag from 'dash-table/derived/cell/resolveFlag';
@@ -34,6 +36,10 @@ const D3_DEFAULT_LOCALE: INumberLocale = {
 const DEFAULT_NULLY = '';
 const DEFAULT_SPECIFIER = '';
 const NULL_SELECTED_CELLS: SelectedCells = [];
+
+const DEFAULT_FILTER_OPTIONS = {
+    case: FilterCase.Sensitive
+};
 
 const data2number = (data?: any) => +data || 0;
 
@@ -63,11 +69,17 @@ const applyDefaultsToColumns = (
     defaultLocale: INumberLocale,
     defaultSort: SortAsNull,
     columns: Columns,
-    editable: boolean
+    editable: boolean,
+    filterOptions: IFilterOptions | undefined
 ) =>
     R.map(column => {
         const c = R.clone(column);
         c.editable = resolveFlag(editable, column.editable);
+        c.filter_options = {
+            ...DEFAULT_FILTER_OPTIONS,
+            ...(filterOptions ?? {}),
+            ...(c.filter_options ?? {})
+        };
         c.sort_as_null = c.sort_as_null || defaultSort;
 
         if (c.type === ColumnType.Numeric && c.format) {
@@ -105,7 +117,8 @@ export default class Sanitizer {
                   locale_format,
                   props.sort_as_null,
                   props.columns,
-                  props.editable
+                  props.editable,
+                  props.filter_options
               )
             : [];
         const data = props.data ?? [];